refactor(api): share provider filter helpers
- Extract JSON API filter helpers into reusable mixins
- Reuse provider filter parsing across overview endpoints
- Clarify compliance overview filter behavior
- Add API changelog entry
